+++ dashboard/src/features/orgs/layout/OrgLayout/ProjectViewWithState.tsx
@@ -1,5 +1,6 @@
+import { useQueryClient } from '@tanstack/react-query';
 import { useRouter } from 'next/router';
-import { type PropsWithChildren, useMemo } from 'react';
+import { type PropsWithChildren, useEffect, useMemo } from 'react';
 import { Alert } from '@/components/ui/v2/Alert';
 import { ApplicationProvisioning } from '@/features/orgs/projects/common/components/ApplicationProvisioning';
 import { ApplicationRestoring } from '@/features/orgs/projects/common/components/ApplicationRestoring';
@@ -15,6 +16,14 @@ function ProjectViewWithState({ children }: PropsWithChildren) {
     query: { appSubdomain },
     route,
   } = useRouter();
+
+  const queryClient = useQueryClient();
+
+  useEffect(() => {
+    return () => {
+      queryClient.clear();
+    };
+  }, [queryClient]);
 
   const { state } = useAppState();
